编辑问题，AnswerAction重构完成

+++ src/main/java/org/gxfj/iknow/service/AnswerServiceImpl.java
@@ -47,7 +47,7 @@ public class AnswerServiceImpl implements AnswerService{
     @Override
     public Map<String, Object> getQuestiontitle(Integer questionId) {
         User user = (User) ActionContext.getContext().getSession().get(ConstantUtil.SESSION_USER);
-        Map<String, Object> result = new HashMap<>(ConstantUtil.MIN_HASH_MAP_NUM);
+        Map<String, Object> result = new HashMap<>(ConstantUtil.RESPONSE_NUM);
         if(user == null){
             result.put(ConstantUtil.JSON_RETURN_CODE_NAME, ConstantUtil.UN_LOGIN);
         }
@@ -283,7 +283,7 @@ public class AnswerServiceImpl implements AnswerService{
     @Override
     public Map<String, Object> adoptAnswer(Integer answerId) {
         User user = (User) ActionContext.getContext().getSession().get("user");
-        Map<String , Object> result = new HashMap<>(ConstantUtil.MIN_HASH_MAP_NUM);
+        Map<String , Object> result = new HashMap<>(ConstantUtil.RESPONSE_NUM);
         if (user != null) {
             if (adoptAnswer(user, answerId)) {
                 //用户已登录，且用户为题主，返回采纳成功
@@ -324,7 +324,7 @@ public class AnswerServiceImpl implements AnswerService{
     @Override
     public Map<String, Object> cancelAnonymous(Integer answerId) {
         User user = (User) ActionContext.getContext().getSession().get("user");
-        Map<String , Object> result = new HashMap<>(ConstantUtil.MIN_HASH_MAP_NUM);
+        Map<String , Object> result = new HashMap<>(ConstantUtil.RESPONSE_NUM);
         Answer answer = answerDAO.getNotDelete(answerId);
         if (user != null) {
             //用户是答主
@@ -360,7 +360,7 @@ public class AnswerServiceImpl implements AnswerService{
     @Override
     public Map<String, Object> getRecommendAnswerForUser(Integer count) {
         User user = (User) ActionContext.getContext().getSession().get("user");
-        Map<String , Object> result = new HashMap<>(ConstantUtil.MIN_HASH_MAP_NUM);
+        Map<String , Object> result = new HashMap<>(ConstantUtil.RESPONSE_NUM);
         Map<String,Object> cUser=new HashMap<>(2);
         if (user != null) {
             result=getRecommendJsonItems(selectRecommendAnswer(user.getId(), count, 0));
@@ -636,7 +636,7 @@ public class AnswerServiceImpl implements AnswerService{
     @Override
     public Map<String, Object> approveAnswer(Integer answerId) {
         User user = (User) ActionContext.getContext().getSession().get("user");
-        Map<String , Object> result = new HashMap<>(ConstantUtil.MIN_HASH_MAP_NUM);
+        Map<String , Object> result = new HashMap<>(ConstantUtil.RESPONSE_NUM);
         if(user == null){
             result.put(ConstantUtil.JSON_RETURN_CODE_NAME, ConstantUtil.UN_LOGIN);
         }
@@ -665,7 +665,7 @@ public class AnswerServiceImpl implements AnswerService{
     @Override
     public Map<String, Object> cancelApprove(Integer answerId) {
         User user = (User) ActionContext.getContext().getSession().get("user");
-        Map<String , Object> result = new HashMap<>(ConstantUtil.MIN_HASH_MAP_NUM);
+        Map<String , Object> result = new HashMap<>(ConstantUtil.RESPONSE_NUM);
         if(user == null){
             result.put(ConstantUtil.JSON_RETURN_CODE_NAME, ConstantUtil.UN_LOGIN);
         }
@@ -686,7 +686,7 @@ public class AnswerServiceImpl implements AnswerService{
     @Override
     public Map<String, Object> oppositionAnswer(Integer answerId) {
         User user = (User) ActionContext.getContext().getSession().get("user");
-        Map<String , Object> result = new HashMap<>(ConstantUtil.MIN_HASH_MAP_NUM);
+        Map<String , Object> result = new HashMap<>(ConstantUtil.RESPONSE_NUM);
         if(user == null){
             result.put(ConstantUtil.JSON_RETURN_CODE_NAME, ConstantUtil.UN_LOGIN);
         }
@@ -709,7 +709,7 @@ public class AnswerServiceImpl implements AnswerService{
     @Override
     public Map<String, Object> cancelOppose(Integer answerId) {
         User user = (User) ActionContext.getContext().getSession().get("user");
-        Map<String , Object> result = new HashMap<>(ConstantUtil.MIN_HASH_MAP_NUM);
+        Map<String , Object> result = new HashMap<>(ConstantUtil.RESPONSE_NUM);
         if(user == null){
             result.put(ConstantUtil.JSON_RETURN_CODE_NAME, ConstantUtil.UN_LOGIN);
         }
@@ -727,24 +727,26 @@ public class AnswerServiceImpl implements AnswerService{
     @Override
     public Map<String, Object> moreRecommendAnswer( Integer count, Integer start) {
         User user = (User) ActionContext.getContext().getSession().get("user");
-        Map<String , Object> result = new HashMap<>(ConstantUtil.MIN_HASH_MAP_NUM);
+        Map<String , Object> result = new HashMap<>(ConstantUtil.RESPONSE_NUM);
+        Map<String, Object> answers = null;
         if (user != null) {
-            if(getRecommendJsonItems(selectRecommendAnswer(user.getId(), count, start))==null) {
+            answers = getRecommendJsonItems(selectRecommendAnswer(user.getId(), count, start));
+            if(answers == null) {
                 result.put(ConstantUtil.JSON_RETURN_CODE_NAME,ConstantUtil.NO_MORE);
             }
             else{
-                result = getRecommendJsonItems(selectRecommendAnswer(user.getId(), count, start));
+                result = answers;
                 result.put(ConstantUtil.JSON_RETURN_CODE_NAME,ConstantUtil.SUCCESS);
             }
 
         } else {
-            if(getRecommendJsonItems(selectRecommendAnswer
-                    (null,ConstantUtil.RECOMMEND_ANSWERS_NUM_PER_TIME,start))==null){
+            answers = getRecommendJsonItems(selectRecommendAnswer
+                    (null,ConstantUtil.RECOMMEND_ANSWERS_NUM_PER_TIME,start));
+            if(answers == null){
                 result.put(ConstantUtil.JSON_RETURN_CODE_NAME,ConstantUtil.NO_MORE);
             }
             else{
-                result = getRecommendJsonItems(selectRecommendAnswer
-                        (null,ConstantUtil.RECOMMEND_ANSWERS_NUM_PER_TIME,start));
+                result = answers;
                 result.put(ConstantUtil.JSON_RETURN_CODE_NAME,ConstantUtil.SUCCESS);
             }
         }
@@ -815,40 +817,65 @@ public class AnswerServiceImpl implements AnswerService{
     }
 
     @Override
-    public boolean deleteAnswer(User user , Integer answerId){
-        Answer answer = answerDAO.get(answerId);
-        if( answer.getQuestionByQuestionId().getAnswerByAdoptId() !=null
-                && answer.getQuestionByQuestionId().getAnswerByAdoptId().getId().equals(answerId)){
-            setQuestionNotAdopt(answer.getQuestionByQuestionId());
-        }
-        for (Comment comment : answer.getCommentsById()) {
-            Collection<Reply> replyCollection = comment.getRepliesById();
+    public Map<String, Object> deleteAnswer(Integer answerId){
+        User user = (User) ActionContext.getContext().getSession().get("user");
+        Map<String , Object> result = new HashMap<>(ConstantUtil.RESPONSE_NUM);
+        if (user == null) {
+            result.put(ConstantUtil.JSON_RETURN_CODE_NAME,ConstantUtil.UN_LOGIN);
+        } else {
+            if(isAnswerer(answerId,user)) {
+                Answer answer = answerDAO.get(answerId);
+                if( answer.getQuestionByQuestionId().getAnswerByAdoptId() !=null
+                        && answer.getQuestionByQuestionId().getAnswerByAdoptId().getId().equals(answerId)){
+                    setQuestionNotAdopt(answer.getQuestionByQuestionId());
+                }
+                for (Comment comment : answer.getCommentsById()) {
+                    Collection<Reply> replyCollection = comment.getRepliesById();
 
-            for (Reply reply : replyCollection) {
-                replyDAO.delete(reply);
+                    for (Reply reply : replyCollection) {
+                        replyDAO.delete(reply);
+                    }
+                    commentDAO.delete(comment);
+                }
+                answerDAO.delete(answer);
+
+                result.put(ConstantUtil.JSON_RETURN_CODE_NAME, ConstantUtil.SUCCESS);
+            } else{
+                result.put(ConstantUtil.JSON_RETURN_CODE_NAME,ConstantUtil.USER_IS_NOT_ANSWERER_TWO);
             }
-            commentDAO.delete(comment);
         }
-        answerDAO.delete(answer);
-
-        return true;
+        return result;
     }
 
     @Override
-    public String getAnswerContent(Integer answerId){
-        Answer answer = answerDAO.get(answerId);
-        return answer.getContentHtml();
+    public Map<String, Object> getAnswerContent(Integer answerId){
+        User user = (User) ActionContext.getContext().getSession().get("user");
+        Map<String , Object> result = new HashMap<>(ConstantUtil.RESPONSE_NUM);
+        if (user == null) {
+            result.put(ConstantUtil.JSON_RETURN_CODE_NAME,ConstantUtil.UN_LOGIN);
+        } else {
+            if(isAnswerer(answerId,user)) {
+                result.put("content",answerDAO.get(answerId).getContentHtml());
+                result.put(ConstantUtil.JSON_RETURN_CODE_NAME, ConstantUtil.SUCCESS);
+            } else{
+                result.put(ConstantUtil.JSON_RETURN_CODE_NAME,ConstantUtil.USER_IS_NOT_ANSWERER_TWO);
+            }
+        }
+        return result;
     }
 
     @Override
-    public boolean updateAnswerContent(Integer answerId, String content){
-        if (!TextVerifyUtil.verifyCompliance(content)) {
-            return false;
+    public Map<String, Object> updateAnswerContent(Integer answerId, String content){
+        Map<String , Object> result = new HashMap<>(ConstantUtil.RESPONSE_NUM);
+        if (TextVerifyUtil.verifyCompliance(content)) {
+            Answer answer = answerDAO.get(answerId);
+            answer.setContentHtml(content);
+            answer.setContentText(HtmlUtil.html2Text(HtmlUtil.changeImgTag(content)));
+            answerDAO.update(answer);
+            result.put(ConstantUtil.JSON_RETURN_CODE_NAME, ConstantUtil.SUCCESS);
+        } else {
+            result.put(ConstantUtil.JSON_RETURN_CODE_NAME, JSON_RESULT_CODE_VERIFY_TEXT_FAIL);
         }
-        Answer answer = answerDAO.get(answerId);
-        answer.setContentHtml(content);
-        answer.setContentText(HtmlUtil.html2Text(HtmlUtil.changeImgTag(content)));
-        answerDAO.update(answer);
-        return true;
+        return result;
     }
 }
